requirements.txt
openai~=1.104.2
python-dotenv~=1.1.1